Organized and cleaned code
