--- HuffmanTree/ht.json
@@ -1,4 +1,0 @@
-{
-    "symbols" : ["", "A", "", "", "B", "", "I", "S"],
-    "freq" : [5, 3, 2, 0, 3, 1, 1]
-}
